DietApp/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.22.0
pydantic>=2.9.0
pandas==2.2.2
numpy>=2.0.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
mistralai>=1.2.0
rapidfuzz

